refactor: event invocation and listerner
