+++ mobile/lib/pages/common/tab_shell.page.dart
@@ -2,6 +2,8 @@ import 'package:auto_route/auto_route.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:immich_mobile/domain/models/timeline.model.dart';
+import 'package:immich_mobile/domain/utils/event_stream.dart';
 import 'package:immich_mobile/extensions/build_context_extensions.dart';
 import 'package:immich_mobile/providers/app_settings.provider.dart';
 import 'package:immich_mobile/providers/backup/drift_backup.provider.dart';
@@ -15,7 +17,6 @@ import 'package:immich_mobile/providers/websocket.provider.dart';
 import 'package:immich_mobile/routing/router.dart';
 import 'package:immich_mobile/services/app_settings.service.dart';
 import 'package:immich_mobile/utils/migration.dart';
-import 'package:immich_mobile/utils/timeline_util.dart';
 
 @RoutePage()
 class TabShellPage extends ConsumerStatefulWidget {
@@ -155,7 +156,7 @@ class _TabShellPageState extends ConsumerState<TabShellPage> {
 void _onNavigationSelected(TabsRouter router, int index, WidgetRef ref) {
   // On Photos page menu tapped
   if (router.activeIndex == 0 && index == 0) {
-    TimelineUtil.scrollToTop();
+    EventStream.shared.emit(const ScrollToTopEvent());
   }
 
   // On Search page tapped
