+++ mobile/lib/presentation/widgets/asset_viewer/top_app_bar.widget.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:immich_mobile/constants/enums.dart';
 import 'package:immich_mobile/domain/models/asset/base_asset.model.dart';
+import 'package:immich_mobile/domain/models/timeline.model.dart';
 import 'package:immich_mobile/domain/utils/event_stream.dart';
 import 'package:immich_mobile/extensions/build_context_extensions.dart';
 import 'package:immich_mobile/presentation/widgets/action_buttons/cast_action_button.widget.dart';
@@ -16,7 +17,6 @@ import 'package:immich_mobile/providers/routes.provider.dart';
 import 'package:immich_mobile/providers/user.provider.dart';
 import 'package:immich_mobile/providers/websocket.provider.dart';
 import 'package:immich_mobile/routing/router.dart';
-import 'package:immich_mobile/utils/timeline_util.dart';
 
 class ViewerTopAppBar extends ConsumerWidget implements PreferredSizeWidget {
   const ViewerTopAppBar({super.key});
@@ -59,7 +59,7 @@ class ViewerTopAppBar extends ConsumerWidget implements PreferredSizeWidget {
           onPressed: () async {
             await context.maybePop();
             await context.navigateTo(const TabShellRoute(children: [MainTimelineRoute()]));
-            TimelineUtil.scrollToDate(asset.createdAt);
+            EventStream.shared.emit(ScrollToDateEvent(asset.createdAt));
           },
           icon: const Icon(Icons.image_search),
           tooltip: 'view_in_timeline',
