+++ mobile/lib/presentation/widgets/memory/memory_bottom_info.widget.dart
@@ -4,8 +4,9 @@ import 'package:auto_route/auto_route.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:immich_mobile/domain/models/memory.model.dart';
+import 'package:immich_mobile/domain/models/timeline.model.dart';
+import 'package:immich_mobile/domain/utils/event_stream.dart';
 import 'package:immich_mobile/routing/router.dart';
-import 'package:immich_mobile/utils/timeline_util.dart';
 
 class DriftMemoryBottomInfo extends StatelessWidget {
   final DriftMemory memory;
@@ -51,7 +52,7 @@ class DriftMemoryBottomInfo extends StatelessWidget {
             onPressed: () {
               context.maybePop();
               context.navigateTo(const TabShellRoute(children: [MainTimelineRoute()]));
-              TimelineUtil.scrollToDate(fileCreatedDate);
+              EventStream.shared.emit(ScrollToDateEvent(fileCreatedDate));
             },
             shape: const CircleBorder(),
             color: Colors.white.withValues(alpha: 0.2),
